requirements.txt
Django==3.2.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
Pillow==9.2.0
pycodestyle==2.9.1
PyJWT==2.4.0
pytz==2022.2.1
sqlparse==0.4.2
toml==0.10.2
whitenoise==6.2.0
django-allauth==0.42.0
dj-rest-auth == 1.1.0
django-allauth==0.42.0
dj-rest-auth == 1.1.0
django_on_heroku
pyotp
django-environ